Implemented run for texture related methods.
