+++ crates/cyancia_brush/src/lib.rs
@@ -3,10 +3,7 @@ use cyancia_assets::AssetAppExt;
 use cyancia_runtime::{Application, plugin::Plugin};
 use cyancia_tools::ToolsAppExt;
 
-use crate::{
-    asset::{BrushPresetSerializer, ImageSerializer},
-    tool::BrushTool,
-};
+use crate::{asset::BrushPresetSerializer, tool::BrushTool};
 
 pub mod asset;
 pub mod browser;
@@ -21,7 +18,6 @@ pub struct BrushPlugin;
 impl Plugin for BrushPlugin {
     fn build(&self, app: &mut Application) {
         app.add_asset_serializer::<BrushPresetSerializer>()
-            .add_asset_serializer::<ImageSerializer>()
             .add_tool_function::<BrushTool>();
     }
 }
